+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * When Steam was not installed, the logs would contain a `warning`-level message.
     This has been demoted to an `info`-level message.
   * GUI: Fixed some inconsistent spacing between elements.
+  * CLI: On Linux, the `wrap` command's `--infer steam` option would fail
+    to find the `SteamAppId` environment variable due to a case mismatch.
 * Changed:
   * GUI: Updated to the latest version of [Iced](https://github.com/iced-rs/iced).
     If the GUI fails to load, Ludusavi will log the error info.